Update to latest Python code
Change histogram binning to use double for width and bounds
to avoid small discrepancies when compared to the apav library
loaded source Python notebook.

Support new implementations and models and allow swtiching
between them in with properties.

+++ Cameca.CustomAnalysis.PeakDetection/HistogramDataProvider.cs
@@ -6,9 +6,9 @@ namespace Cameca.CustomAnalysis.PeakDetection;
 
 internal class HistogramDataProvider : IPyObjectProvider
 {
-    private readonly ReadOnlyMemory<float> data;
+    private readonly ReadOnlyMemory<double> data;
 
-    public HistogramDataProvider(ReadOnlyMemory<float> data)
+    public HistogramDataProvider(ReadOnlyMemory<double> data)
     {
         this.data = data;
     }
